+++ internal/comby/comby.go
@@ -13,36 +13,16 @@ import (
 const combyPath = "comby"
 const numWorkers = 8
 
-type Range struct {
-	Offset int `json:"offset"`
-	Line   int `json:"line"`
-	Column int `json:"column"`
-}
-
 // {"uri":"/private/tmp/rrrr/doc.go","matches":[{"range":{"start":{"offset":215,"line":1,"column":216},"end":{"offset":222,"line":1,"column":223}},"environment":[],"matched":"package"}]}
-type Match struct {
-	URI     string  `json:"uri"`
-	Matches []Range `json:"matches"`
-	Matched string  `json:"matched"`
-}
-
-type Diff struct {
-	URI  string `json:"uri"`
-	Diff string `json:"diff"`
-}
 
 // A result is either a match result or a diff result. These members are
 // mutually exclusive, but are bundled together so that we can share the
 // unmarshalling code.
-type Result struct {
-	Matches *[]Match
-	Diffs   *[]Diff
-}
 
 func exists() (_ int, err error) {
 	_, err = exec.LookPath("comby")
 	if err != nil {
-		return -1, errors.New("comby is not installed on the PATH. Try running 'bash <(curl -sL get.comby.dev)'.")
+		return -1, errors.New("comby is not installed on the PATH. Try running 'bash <(curl -sL get.comby.dev)'")
 	}
 	return 0, nil
 }
